Calculate table rows dynamically
To prevent vertical scrolling issues on short touch-screen
devices, the table's height is reduced to the point where
it will be visible in entirety within the shown screen.

To do this, we calculate the number of rows that should be
shown dynamically, by assuming a minimum height of portrait
iPad (768px), and adding rows for however much the actual
screen is taller than that minimum.

The utility function that calculates this takes the average
row height and minimum number of rows as parameters, so it
can be reused for tables that have differing characteristics.