Switch to MapshedJob for Analyze, TR-55
The MapshedJob module of the geoprocessing service is flexible and
supports a number of operations given the right input JSON. It can
perform the RasterGroupedCount operation without requiring special
code to do it. Thus, we move away from SummaryJob.

The input JSON is very similar to that used for existing MapShed
jobs. The "bottom half" of the histogram function, used to process
the retrieved results, is amended to perform remapping of soil
codes, which was previously done within SummaryJob, since the new
MapshedJob does not support reassigning values. The results are
the compiled into an array, just as they were previously, allowing
the consumers of `histogram_finish` to be unaffected by this.