Fix bug in census counting
Previously we would overwrite existing values with new counts,
assuming that each key would have exactly one instance thus it
being a non-destructive operation. Unfortunately, as a result
of the soil re-mapping, there are duplicate keys in the dict:

    c:developed_med: 10
    c:developed_med: 8411

In this case, the later value would be used, and the results
would differ based on the order of values in the input array.
This is bad because of two reasons:

 * The input values are a set, not a list, so the order should
   not matter.
 * We lose data when we overwrite it.

Now we sum the values instead, to ensure that the final counts
are correct.