Refactor TR-55 chain to use geoprocessing tasks
Previously the TR-55 Celery chain was using custom methods
for running Spark JobServer, and a severely multi-layered
set of methods to process the results.

Now, we use the standard geoprocessing start and finish
methods to run Spark JobServer. A simple nlcd_soil_census
method is used to calculate the census from those results.