Split analyze into land and soil endpoints
Previously there was a single endpoint for getting land and
soil surveys for an area of interest, to be used in Analyze
view of the front-end. This reused the geoprocessing design
for TR-55 for convenience. Unfortunately, this was a costly
decision, for TR-55 geoprocessing service overlaid land and
soil values and returned a combined result. This result was
useful for TR-55, but not for Analyze, which only needs the
discrete counts for each.

By switching to independent endpoints, we make Analyze much
faster and simpler. As a consequence, the values can now no
longer be shared with TR-55, causing it to start a new geo-
processing chain when it is triggered. However, this is an
acceptable compromise, as the user may choose another model
such as MapShed to run after Analyze, in which case the old
values would be discarded anyway.

In this commit, we:

 * Add nlcd and soil blocks to geoprocessing JSON config
 * Replace /analyze with /analyze/land and /analyze/soil
 * Add two new views, which use the standardized pattern
   of geoprocessing.start, geoprocessing.finish, and one
   post-processing function each.
 * Add post-processing functions for both land and soil
   survey calculation. The functions are very similar, but
   kept separate to avoid the multi-level layering of the
   arrangement before.
 * Remove unused functions from tasks and geoprocessing
 * Update front-end task runners to point to new endpoints
 * Update front-end view to correctly find land task runner
 * Remove adding census to the project model in the front-
   end. We could previously do this because the /analyze
   endpoint would return a census featuring combined values
   of land and soil, as well as independent surveys of both.
   The surveys would be used in the Analyze view, and census
   would be passed to Model as pre-cached geoprocessing.
   Since the new endpoints send back only their respective
   surveys, and not the combined census, we remove that
   functionality on the front-end.
   Now the behavior is akin to the user proceeding from
   Analyze to Model without waiting for the analysis to
   finish.