Parse input into area of interest and WKAoI id
Previously we would always send the area of interest as JSON in
the model input. Now, we’d like to add support for sending only
the WKAoI id, and have it picked up from the database.

To this end, the parse_input function takes either a dict or a
JSON string, and inspects if it carries either a ‘wkaoi’ or an
‘area_of_interest’ key. If the former is found, we look it up
in the database and return the shape from there, along with the
given WKAoI id itself. If the latter is found, we convert it to
a single ring multipolygon (to account for the multiring multi-
polygons sometimes generated via RWD), and return None as the
WKAoI id.

In the front-end, analyze requests have been updated to send in
an ‘area_of_interest’ key under ‘analyze_input’, following the
style used for TR-55 ‘model_input’ and MapShed ‘mapshed_input’.