Add WKAoI support to geoprocessing tasks
geoprocessing.start now takes an additional optional parameter for
well known areas of interest. If it is not specified, a standard
geoprocessing operation is run same as before.

When it is specified, the operation name is combined with the
wkaoi to form a cache key, used to see if a cached value for this
operation and input already exist. If found, those values are
passed to geoprocessing.finish immediately, and nothing is sent to
Spark JobServer.

If not found, a Spark JobServer job is begun, and the key is given
to geoprocessing.finish. If the job completes successfully, the
results are stored in the cache with the given key before being
passed along.