Use WKAoI for cached geoprocessing
Update Analyze, TR-55 and MapShed tasks to use the new WKAoI
support in their implementations. For Analyze and MapShed, the
WKAoI is used immediately and always. For TR-55, we are careful
to only WKAoI has no modifications applied to it, lest we cache
results of a different set of inputs to the wrong name.