Add caching support to front-end
Now that caching is in place in the back-end, we update the front-
end to take advantage of it. When selecting a well known area of
interest, in addition to its id and shape, we also make note of
the table name. The table name and id are combined to make a well
known area of interest key ‘wkaoi’, and is sent to the back-end
instead of a full area of interest. In case the value is not
cached, it will be cached after the first use. Otherwise, we get
the value back in a single poll.

The behavior for other kinds of areas of interest: user drawn,
RWD, and user uploaded, remain the same as before. These are not
cached.