Namespace cache key with geop_ prefix
In the ADR-002, we had declared that the geoprocessing cache
key would start with the 'geop_' prefix, for ease of deletion
if the cache needs to be emptied without removing other cached
items in redis, and to prevent accidental name collisions.

This had originally been missed, and is added now.