Set BiG-CZ mode using query param or host
Previously we were setting BiG-CZ specific settings based on
an environment variable. However, since in live deployment both
MMW and BiG-CZ will be running off the same (virtual) servers,
we cannot rely on environment variables to distinguish them.

This approach looks for a ?bigcz=true query string. If found, it
sets a session variable which turns on all relevant switches.
These settings remain until the user’s session expires, or they
pass in ?bigcz=false which turns them off.

In addition, we also inspect the request’s HTTP_HOST header, and
if it matches the BiG-CZ pattern we turn these settings on. If
it matches the MMW pattern we turn these settings off.