Set BiG-CZ mode using query param or host
Previously we were setting BiG-CZ specific settings based on
an environment variable. However, since in live deployment both
MMW and BiG-CZ will be running off the same (virtual) servers,
we cannot rely on environment variables to distinguish them.

This approach looks for a ?bigcz=true query parameter, or the
Host header of the request coming from [*.]portal.bigcz.org.
If the request is coming from that host, or that query parameter
is passed, we enable BiG-CZ settings. Host inspection will be
used in production, and query parameter is to support development.