Fetch CUAHSI services for whole world
Fixes a bug which prevents sites from joining to services, depending on
the size of the AOI. CUAHSI may return results for `GetSitesInBox2` but
may not return results for `GetServicesInBox2` using the same bounding box.
This could be due to using a "contains" rather than "overlaps" intersection
query. The workaround is to fetch services for the entire world, which
is a relatively quick operation (sub second).