Refactor bounds to be supplied from view
Previously the bounds were calculated within the model. Since
we need to calculate the area of the bounds and act on it in
the view (to protect against large queries), we refactor the
model method to accept bounds as a paramter rather than it
calculating the value locally.

This will also make it easier to switch to AoI bounds in the
future.