Remove search bar when project page is launched
Explicitly remove the search bar when the project page is launched. Some
pages remove this element when the page is navigated away from, but this
technique does not work for the splash page. I believe this is because
it is as the root url and does not have a route name.

Refs #1912