Filter results to those intersecting AoI
If the results do not have any geometry, include them. This is
necessary for Hydroshare which has none. But, if there is a
geometric component, filter out those that do not intersect the
area of interest. This will remove those entries that fall within
the bounding box of the area of interest, but not the actual
area of interest itself.