Fix scenario share links
Previously the scenario share modal would show the current URL of
the page. This was okay when the scenarios were tabs, and only the
active scenario had a context menu. However, since now any scenario
can have context menu, we needed to update the share modal to show
the right link, even when the selected scenario is not currently
active on the page.