Use bootstrap-table export extension for CSV
Replace the custom CSV download with the built-in export functionality
of bootstrap-table. CSVs will now match their corresponding table in the
UI exactly.

A slight hack is required to render tables with pagination. An
additional table with the full dataset is rendered beneath the
paginated table. It's then used to generated the CSV and then removed
from the DOM.

Refs #2032