Dynamic serialization of resource lists
In anticipation of providing a serializer at runtime for the
various clients, we change the ResourceListSerializer.results
field to be dynamic. If given a `serializer` in the context
object, we will use that to serialize the results, falling
back to the generic serializer.

I'm not sure if this is the most performant way of doing it,
but we can focus on optimizations at a later stage.

See:

 * http://www.django-rest-framework.org/api-guide/fields/#serializermethodfield
 * http://www.django-rest-framework.org/api-guide/serializers/#including-extra-context

:hammer: