Support array of catalog results from back-end
In order to support the use case of querying without specifying a
catalog and returning multiple catalogs in the result, we switch
the back-end to return an array of results instead of a singleton.
The front-end is correspondingly updated to find the specific set
in the response that matches the catalog it was initialized with
and display accordingly.