Centralize Max Area setting
Previously, the max allowed areas for MMW and BiG-CZ were defined
on an ad-hoc basis. Now we centralize them, and expose them via
the settings.max_area attribute, to be reused everywhere in the
app.

Also, now the area limit of CUAHSI is applied to all of BiG-CZ,
since a user may not know their AoI will not work with CUAHSI
until it's too late to change the AoI. Thus, we warn them earlier
in the process.