Enable toggling between analyze and BigCZ results
Allow the user to switch between the AoI analyze results, and the BigCZ
search page. This implementation is based on the analyze and modeling
tabs of MMW. Some views, templates, and CSS rules from that implementation
were duplicated and adjusted rather than abstracting the MMW
implementation and making it reusable.

Refs #2095