Add sync methods for geoprocessing
We add a task `run` and a helper method `geoprocess`. The `run`
task converts the input into the desired format, and `geoprocess`
communicates with the geoprocessing service and returns results.

`run` is a combination of `start` and `finish`: it checks whether
a result is cacheable and cached or not, and if so returns that.
Otherwise it runs `geoprocess`.

`geoprocess` is similar to `sjs_submit` in the sense that it is
POSTing to an endpoint. Unlike `sjs_submit`, which gets back a job
id, `geoprocess` receives the actual results and returns them.

`run` is designed to replace `start` and `finish` tasks in Celery
chains. So if a previous celery chain was:

    chain(geoprocessing.start.s(data),
          geoprocessing.finish.s(),
          mytasks.process_results.s())

It will now be:

    chain(geoprocessing.run.s(data),
          mytasks.process_results.s())