Make y-axis tick precision dynamic
In case of a stacked chart, or when the minimum y-value is
more than 1, we use the 0.1f format which is 1 decimal place.
For when the value is less than 1, we calculate how deep is
the most significant digit of the smallest non-zero y-value,
and set that as the tick precision.

See https://github.com/d3/d3-format for details.