Export ModalBaseView
In order to reuse the modal setup for the Compare View window,
the ModalBaseView is now exported in addition to the other
"finished" modal views.

It is also switched to be a LayoutView instead of an ItemView
so that the Compare View Modal can have regions. Otherwise,
the LayoutView is identical to the ItemView and should not
affect the behavior of other modals that extend from this.