Modalify the Compare View Window
Instead of being a simple LayoutView, the Compare View Window
is now a modal. This darkens the background of the page, and
disables scrolling outside the modal. Clicking outside it will
not close it, but pressing the X button in the top right, or
pressing Esc on the keyboard will close the modal.

Even though this is a modal, there are no div's with .modal-body
or .modal-header classes. Thus, the contents do not look like
the other modals, and we are not affected by the .modal-basic
class inherited from ModalBaseView.