Remove custom celery routing and exchange
Due to the implementation of async SJS requests, we established a
mechninism for directly establishing routes that would stay on a single
worker.  This led to complex code and also high latency when we pinged
each worker to determine if they were available.  With recent changes
removing SJS in favor of a synchronous geoprocessing call, we can now
cull these custom routes and exchanges, both simplifying the code and
increasing the speed of submitting a new job.