Add support for search options in front-end
Add `has_filters` and `options` attributes to Catalog model. If
`has_filters` is true, then we show the UI section for filters,
and wire up some handlers.

When a new Catalog is created, we can optionally give it a
collection of SearchOptions, each with an `id` and an `active`
state. When making search requests, all Options that are
`active` will be joined together into a comma-separated
list and sent as `options=a,b,c` to the sever.