Pass full BIGCZ result models to map
Previously, just the geometry for the selected (both the active and
detail) result was passed to the map. In order to faciliate more complex
interaction between the map and the sidebar, the full result model is
now passed to the map. This will allow functions within the map view to
read and set properties of the result.

Refs #2181