Update Compare Window model when polling
We introduce a new `polling` boolean attribute on the Compare
Window model, false by default, which reflects whether the window
is currently in a state of polling or not. This is contingent on
any of the scenarios having any results that are polling.

To do this, we first collect all the `polling` attributes in each
of the result in each scenario, and then flatten it to a single
array. Then we use `Array.some` which takes a function and returns
true if it evaluates to a truthy value for any of the values.
Since the values are already boolean, we simply use `_.identity`
as the evaluating function.