Fade results when polling
When the Window Model has `polling` set to true, the sections
region containing charts or tables will appear faded. It will
come back into saturation once the polling stops. This change
is animated in the same manner as all other animations in the
Compare Window.

This does not impede user interactivity. They can still hover
over the charts and select text in tables.