Add spinner on poll
Add a spinner on the top middle, left of the precipitation slider,
to be shown when the scenarios are polling.

This uses a new spinner which is sleeker and faster than the usual
fa-circle-o-notch used elsewhere in the app. Eventually, all of those
should be replaced with this.

squash! Add spinner on poll

The spinner is now in the middle of the screen, rather than the top-right.