Set and unset WKAoI wherever AoI was changed
Since now the WKAoI field is saved in the model, we need to
update the map's wellKnownAreaOfInterest field accordingly.
In addition, since the AnalyzeCollection is tailored to a
specific AoI on creation, when exiting the Modeling view we
destroy it, so that it may be made again with the correct
WKAoI (if any).