Correct caching of parametrizable operations
Since `ppt` and `tmean` can be run for a number of different
rasters, we cannot cache them with just the operation name. A new
optional `cache_key` parameter is added for caching that appends
a unique identifier per month to the caching key, ensuring that
all parameterized calls of these operations are cached separately.