Upgrade geoprocessing to support area pixels
Previously, for low resolution rasters with very large pixels,
when paired with small areas of interest, we would get 0 values.
This was because the pixels were treated as points, and if the
centroid of the pixel was outside the area of interest, there
would be no intersection.

By adding a `pixelIsArea: true` parameter, we treat those pixels
as polygons instead of points, and can calculate the intersection
even with small areas of interest. The Precipitation and Temperature
rasters are prime use cases for this, and their usage is corrected
with this new setting.