Make popover text selectable
Previously, with 'focus', the popover would disappear after a
click. With 'click', it stays around until explicitly closed,
either by the close button within the popover, or by re-clicking
the info icon that opened it. This allows the user to select
text in the popover, and in the case of URLs, paste them into
new tabs to visit the relevant sites.