Serialize CUAHSI / WDC responses
The suds library interacts with the CUAHSI SOAP API by reading
their WSDL and creating dynamic classes on the fly for requests
and responses. Since these classes are created at runtime, they
are not serializable, since pickle (or json or anything else)
will not know how to deserialize them, as their class definitions
will not be available.

In preparation for caching these values, they must be made
serializable. To this end, we convert the results into Python
Dicts, which are serializable. This uses the suds `asdict` method,
which converts a suds Object to a Python Dict. Arrays are
converted to Lists, literals used immediately, and sub-Objects
are converted recursively to Dicts.