Add `activeVar` to Climate Analyze model
The user will need to be able to switch between two types
of charts: Precipitation and Temperature. This new `activeVar`
variable, which will be one of 'ppt' or 'tmean', will set
which one is shown to the user.

This is modeled after the similar variable used in GWLF-E
runoff views, which also use a dropdown in conjunction with
a line chart.