Don't apply error handlers on chains with groups
Previously we could attach an error handler to any Celery chain
and have it catch an error in any task of the chain. This did
not work with groups, which were ignored, so we had to apply
the error handler manually to every task in the group.

With the latest version of Celery, the groups are no longer
ignored. Instead, we get a runtime error saying that chains
with groups cannot have error handlers attached to them. Thus,
we remove the error handlers on chains that have groups in
them.

Since all the group tasks have individual error handler tasks
already, and the other tasks are simple data transformations
in Python, this should not affect the output.