Remove extraneous buffer tasks for chained chords
The previous version of Celery had a bug which did
not allow a chord to be the antepenultimate task
in a chain. To get around this, we added extraneous
tasks as buffer to get it to work correctly.

The latest version of Celery no longer has this bug,
so these tasks are collapsed into their successors.
This will reduce Celery overhead as we run fewer tasks.

See https://github.com/celery/celery/issues/3191