Switch to new errback format
Previously, a Celery error callback was only given the UUID
of the failing job. To retrieve additional information about
it, we had to query the app with it. To get the app reference,
the task had to be bound to the app.

With new Celery, there is an issue with using bound tasks for
error handling. See https://github.com/celery/celery/issues/3723

Fortunately, in the new error callback format in Celery 4,
the handler is given a request, exception, and traceback as
arguments. This obviates the need for an app reference, since
these were the values we were fetching from it. Now we use
them directly.