Reduce geoprocessing max retries
Previously, since Spark JobServer would take a long time and
much resources to accomplish, sometimes to the point of not
accepting new requests, we had to retry a large number of
times in hopes of making it in eventually.

Since the new geoprocessing service is quite fast, and is
built on Akka which more readily accepts new requests even
when running processing on background threads, we don't
need to retry as much. By reducing excessive retries, we
make sure to not clog up the Celery worker capacity.