Fix gis_data serialization
We recently changed the `ProjectSerializer` to use a `JsonField` serializer for
the `gis_data` field to reflect that fact that this data is now returned from
the geoprocessing service as an actual JSON object rather than a string
containing JSON data. Because this field serializer change was not also made to
the `ProjectUpdateSerializer` we ended up saving the string representation of a
Python dict into the database rather than a JSON string. This caused the model
to crash when attempting to load a previously saved project.

Adding a plain `JsonField()` to the `ProjectUpdateSerializer` caused test
failures that were resolved by adding attributes signifying that it is a
nullable field, which matches the underlying model. I changed both usages of
`JsonField` for consistency.

I tried to make `ProjectUpdateSerializer` inherit from `ProjectSerializer` but
it caused problems with scenario serialization.

I took advantage of the fact that `ast.literal_eval` works for both JSON and
Python dict text formats to write a simple corrective data migration.