Reverse Gridded Services Filter
Previously, the filter was called "Gridded Services" and disabled
by default. If it were turned on, a search option named "gridded"
would be sent, to include gridded services.

However, this is not how most search filters behave. Intuitively,
each filter reduces the number of results, and disabling should
increase the number of results.

To better match this intuition, the filter is now reversed to be
named "Exclude Gridded Services". When it is turned on, the
gridded services will be excluded, and turning it off will include
them, thus expanding the number of results.

This filter is now turned on by default, to maintain current
behavior of not including gridded services unless explicitly
requested by the user.