Different views for each catalog
Previously we were using the same Marionette view with
differing templates for each catalog. Now that we are
trying to add custom events and sub-views, this approach
no longer suffices.

We make a base view which has open / close functionality
built in, and extend it for each catalog, supplying the
template in the child view. Instead of a dictionary of
templates, CATALOG_RESULT_DETAILS_VIEWS now holds the
corresponding view.