Fetch variable dates, then fetch values
Different sites and variables have values only for certain
ranges. Querying for ranges outside of those lead to errors.

Now we first query the /details endpoint to get the units and
begin and end date for each variable, then validate the given
search dates to be within those (and use sensible defaults of
1 week or entire date range, whichever is shorter, if no from
and to dates are specified).

Furthermore, we only search each site once. If the values are
successfully retrieved, they will be shown immediately the
next time the user opens the details page for that site. If
the value retrieval errors out for some reason, we retry that
every time the details page is opened.