Return more details for Variables
Previously, a CUAHSI search would have a `concept_keywords` array
of strings, corresponding to each variable. Now that we want to
fetch more detailed values, we switch to having an array of
`variables`, each of which in addition to `id`, `name` (that
tends to be longer, more like a description than name), and
`concept_keyword`, has a `site` and `wsdl` which will be used
to fetch values for that variable in a given timespan.

This array is sorted by `concept_keyword`, which will also be
used as a display label in the UI.