Add `details` and `values` endpoints
For fetching CUAHSI variable values, we add two endpoints:

`details` expects a `wsdl` and `site` parameter, and uses it to
fetch the WaterML siteinfo object for the given site from the
given wsdl url. This siteinfo object contains details for the
variables in the site, including units, and the time ranges for
which each variable has values. This information is important,
because if we query for `values` with a time outside this range,
the endpoint crashes.

`values` expects a `wsdl`, `site`, `variable`, `from_date` and
`to_date`, and uses it to fetch the values of that variable at
that site in the given time range, and returns the WaterML value
object.

The expected use is one call to `details` for a site, to fetch
units and time ranges for each variable, and then a call to
`values` for each variable in the site, fetching its values for
the given time range.

---

Notes:

1. The WaterML objects themselves are complex and elaborate, with
   many nested keys and arrays. Fortunately, we don't need to
   write our own serializers for them, because that is all taken
   care of within Ulmo. Furthermore, because of WaterML being an
   underlying standard, we can reasonably expect the data to
   follow a certain shape, and exercise this assumption all the
   way on the client side, without having to verify it on the
   server.

2. Ulmo uses suds under the hood to make SOAP requests, and by
   default uses the suds cache, which is distinct from the redis
   cache used elsewhere in the app (for Geoprocessing, etc). I
   have not looked into changing this default behavior, leaving it
   to a future investigation should the need arise.