Fetch CUAHSI Values for Result
For a given CUAHSI Result, we add a `fetchCuahsiValues` function
that first fetches the /details/ for the result, then fetches
/values/ for each variable in the result. If either fetch fails,
we set an `error` field to true. A `fetching` field is added to
track when the fetch completes.

When a CUAHSI Result page is loaded, each result has some fields
of every variable filled out: id, name, concept_keyword, wsdl, and
site. Before we can fetch the variable values, we need begin_date
and end_date, thus we call /details/ to fill those values. Once
that succeeds, the actual values can be fetched.

`units` are populated both by /details/ and /values/. In most
cases they agree, but in some cases they differ. The response
from /values/ always overrides /details/, and is considered
more accurate.