Include geometry for HydroShare results
Previously the HydroShare API did not send any geometries
in its responses. Now, the "coverages" key is an array of
(at least) two types of values: period and box, for time
and geography respectively.

We translate the box coverage into a geometry if it is
included. It behaves identically to CINERGI in the UI,
with respect to multiple overlapping popups and polygons
that exceed the map bounds.