Rearchitect WDC Detail View for refreshes
The main WDC Detail view is a Marionette.LayoutView. Previously,
it featured a number of inline elements that needed to be updated
on model changes, such as the fetching spinner, the error icon,
and the most recent result date. These were siblings to other
Marionette.ItemViews, like the Table View and the Chart View.

Unfortunately, when LayoutViews are rendered, all their child
views are detached, destroyed, recreated, and reattached. This
leads to unnecessary revision of the Table View which is updated
separately, and can cause scroll jumps in some cases because the
entire view is re-rendering.

To get around this, we follow the recommended Marionette pattern
of never re-rendering the LayoutView. The static parts of the
LayoutView remain there, which are only rendered once when the
detail view is loaded. We split the rest into two additional
views: a Status which has the error and fetching icons, and a
Switcher which has the most recent result date, and the table /
chart switcher buttons.

All the child ItemViews listen to the same model fields on their
own and update accordingly.