Add filter.isDefault function
In most cases, filters will have default falsy values, i.e.
they will not be turned on. However, in some cases, such as
WDC's "Exclude Gridded" filter, they will be turned on by
default.

In order to see if the filter's current value is different
from its default value, we add an unimplemented `.isDefault`
function to the FilterModel, with the expectation that it
will be implemented by all child models.

It is implemented in the DateFilter as an _.isEmpty check on
both dates, and in SearchOption as a comparison with the
default active value.

It is also implemented in the FilterCollection to check if
all the filters in the collection satisfy their own isDefault.