Add begin_date, end_date to HydroShare Search Results
These fields are populated from the coverages[type=period] field
in the API results. When they are not present, we return None.

Custom model and serializer were added to accommodate these
fields. They are named the same as Cuahsi's, so if in the future
we consolidate them into the parent Resource class the public
API will not change.