Cache HydroShare API calls for fast filter switch
Previously we would make a new request for each HydroShare API
call. With the addition of the "Exclude Private Resources" filter,
which would add a few items to an existing set of results, making
the user wait for a whole new request is frustrating. To alleviate
this, we cache the API results for 30 minutes. This makes filter
switching very performant.

The 30 minute period is arbitrarily chosen. It may be fine tuned
in the future.

The CUAHSI cache key has been updated to reflect the catalog.