Add UserProfile model related to User and created on login
Following the pattern of `ItsiUser` I have created a new model related to `User`
to hold additional, optional information.

- All additional user profile fields are optional.
- `user_type` is selected from a defined list of choices and defaults to
  "Unspecified."
- `country` is selected from a defined list of choices and defaults to "US."

We will be letting users skip entering this profile information so I have added
a `was_skipped` and `is_complete` flags to the `UserProfile` model and return
their values from the login endpoint so that client code can present the proper
prompts for the information.