Hide user endpoints in Swagger API documentation
The user namespace is added to the user URLs, which, because of our
Swagger configuration, prevents the URLs from showing up in the
documentation.

This was originally added in #2227, but it broke compatibility with
ITSI and was removed in #2377.

The namespace is added here again along with a fix to the ITSI login view.
When calling reverse on a URL which is included in a namespace, the
namespace must also be added to the reverse parameter. See
https://stackoverflow.com/a/38390178/217378.

Refs #2401