Show blue dot if profile form hasn't been filled out
If the user hasn't filled out the profile form yet,
show a blue dot next to their name in the app header.

If the user has just skipped the form for that login
session, show a popover on the dot explaining they
can click the dot to fill the form out whenever
they're ready (it links to the account profile page).

To deal with the timing of when user fetches happen
wrt to showing the profile form, use the profile
response to set attributes on the user model so
another user fetch doesn't have to be done.