Send profile status in GET login and POST profile responses
We were sending the status of the profile (skipped + complete)
only when posting a login. We'll need this information also
for users with an active session so we can display their blue
dot if they need to fill out the profile.

To avoid another GET request, also provide the profile status
information in the response of the /profile request