Use serverResults field for saving CUAHSI results
In preparation for performing client-side free text search on
CUAHSI results, where we will get all results from the server
once and then filter them on the client side.

For the CUAHSI catalog, instead of saving the results to the
`results` field directly, we instead use the `serverResults`
field. Currently, all the results are copied over to the
`results` field, but in coming commits we will add filtering.