Free Text Search
A textFilter method is added to Result model that returns truthy
if a given query is found in any of the CUAHSI fields. The
Results collection also has an identically named method, that
filters the collection to those results that satisfy the query.

When the first search completes, we filter the serverResults by
the given query and include only those in the results field.

On subsequent searches, we check to see if the geometry has
changed or if only the query has changed. Since we fetch all
results for a given geometry in the first search, we don't
need to do other server searches until the geometry changes.

If the geometry is the same, and only the text has changed,
then we re-run the filter on the saved serverResults, and
re-populate results with the filtered items.