Make HydroShare dates timezone aware
Some of the dates in the API response were timezone aware,
while others were not. Since DATE_MIN was defined without
timezone awareness, it could only be compared to similarly
naive datetimes.

Using Django's make_aware utility, DATE_MIN is initialized
with a timezone. Values coming in from the API are checked
for timezone awareness, and are made aware if needed.

This ensures that no information is lost, and comparisons
are made between like objects.