Download results as JSON
Assemble all displayed results' JSON, and download it to a file.
The file will be named in the fashion:

    bigcz-{catalog}-{query}-{date-time-stamp}.json

Note that the format of the JSON is that of the Result Model, and
not of the values that came in from the API. Thus, it will have
some extraneous fields like loading and error. But in the interest
of a efficiency and minimalism, this is tolerable.