Clone reference to gis_data prior to mutating
A recent change to serialization and storage data type on the client
made gis_data, once a string, an object.  As an object, the reference
stored by the project model and in the mergedGisData var were the same,
thus persisting any subsequent changes back to the model, unintentially.
Doing a deepClone of the gis_data before updating values prevents the
core project gis_data from being changed, and having that change shared
across scenarios.