Use full day for single date results
Some results, instead of having a range of dates for their values,
have a single date instead. Previously we would send the same date
as start and end for the range in the API call, which excluded any
data not recorded at midnight of that date.

In cases when the start and end dates are the same, we now add one
day to the end date and ensure data collected any time during that
day is retrieved.