Adjust analyze tab height in tabbed view
When the Analyze window is in a tab, when both analyze
and data are available as tabs, its contents weren't
scrolling correctly.

By calculating the height and taking into account other
items on the page, we ensure that the container's height
is set in a way that allows all content to scroll and
be visible.