Delete fetchPromise on results fetch
When a filter changes, all results for a catalog are refetched.
It is possible that a filter may include results that were
already included. In this case, there isn't a new instance of
the result, but the existing instance is refreshed. When this
happens, its CuahsiVariables collection is reset, with no detail
values.

Previously, if a result details page had been viewed, it had
fetched the detail values for said CuahsiVariables, and cached
them, using a fetchPromise that would only be resolved once.
After fetching them for the first time, we simply returned the
fetchPromise so they would not be fetched again.

When the values are reset, the actual detail values are gone,
but the cached fetchPromise remains, thus preventing values from
being fetched anew. By deleting the fetchPromise when receiving
new values, we ensure that when a detail page is opened for a
previously fetched and repeatedly filtered result, the values
for its CuahsiVariables are refetched.